implementation of vcd-indexer
Implemented vcd-indexer. This could be much, much more interesting.
However, now it is a monolithic piece of code that shows a result only
in the end. To do: make it async (really async) so that it is possible
to send messages from it and inform the user about the current status of
the operations.

Signed-off-by: Andrea Calabrese <andrea1995.c@live.com>